Separate periodic checkpoint responsibilities
